+++ spec/spec_helper.rb
@@ -1,2 +1,3 @@
 $LOAD_PATH.unshift File.expand_path('../../lib', __FILE__)
-require 'acmesmith/command'
+require 'acmesmith/config'
+RSpec::Expectations.configuration.on_potential_false_positives = :nothing
